+++ src/CouchDB.Driver/Types/CouchDocument.cs
@@ -38,6 +38,15 @@ namespace CouchDB.Driver.Types
         [JsonProperty("rev", NullValueHandling = NullValueHandling.Ignore)]
         private string RevOther { set => Rev = value; }
 
+        [JsonIgnore]
+        public bool Deleted { get; private set; }
+        [DataMember]
+        [JsonProperty("_deleted", NullValueHandling = NullValueHandling.Ignore)]
+        private bool DeletedOther
+        {
+            set => Deleted = value;
+        }
+
         [DataMember]
         [JsonProperty("_conflicts")]
         private readonly List<string> _conflicts;