+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * **Bulk Delete**: Adds support to replication ([#171](https://github.com/matteobortolazzo/couchdb-net/issues/171))
 * **Revision Support**: Support for revisions in add and update ([#170](https://github.com/matteobortolazzo/couchdb-net/pull/170))
+* **Deleted Flag**: Added deleted flag on document ([#154](https://github.com/matteobortolazzo/couchdb-net/pull/154))
 
 ## Bug Fixes